Implement "DisplayAllTasks" function
